requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0
aiosqlite>=0.19.0
jinja2>=3.1
python-dotenv>=1.0
httpx>=0.25.0
sse-starlette>=1.6.0
openpyxl>=3.1.0
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=4.9
thefuzz>=0.20
python-multipart>=0.0.6
aiofiles>=23.0
